--- jsconfig.json
@@ -1,7 +1,0 @@
-{
-    "include": [
-        "./node_modules/cypress",
-        "cypress/**/*.js",
-        "cypress/**/**/*.js"
-    ]
-}
